backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
distro==1.9.0
Django==5.0.3
django-extensions==3.2.3
django-filter==23.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
inflection==0.5.1
openai==1.13.3
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
